fix: two real bugs in the GitHub-dump/scheduler paths, plus latent fixes + tests
Surfaced by the workbench clean-code review (0%-coverage prod paths):
- git.py merge_csv_files: passed CSV *content* where read_csv wants a path/buffer
  -> always raised, swallowed -> every .csv GitHub dump silently corrupted. Now
  passes the StringIO/BytesIO buffers directly. Regression test: tests/test_git.py.
- schedule.py import_data: root.rmdir() on an os.walk str -> AttributeError, crashed
  the cron on any empty dir. Now Path(root).rmdir(). Regression test: tests/test_schedule.py.
- normalize_data.closest_hour: ValueError at 23:30+ (hour==24) -> round via timedelta (+test).
- feature_imputation: in-place fillna was a silent no-op under pandas 3.0 CoW -> reassign.
Clean-code: environment exit(-1)->SystemExit; git.py import-frozen default arg->None;
print()->logging; type() is ->isinstance; Series()->Series(dtype). 124 tests green, black-clean.

+++ src/processing/forecast_data.py
@@ -171,7 +171,7 @@ def recursive_forecast(
         DATA_PROCESSED_PATH / city_name / sensor_id, index_col="time"
     )
     if len(dataframe.index) == 0:
-        return Series()
+        return Series(dtype="float64")
 
     dataframe = dataframe.loc[datetime.now() - timedelta(weeks=52) : datetime.now()]
     target = dataframe[pollutant].tail(lags * 2 + 1)
